+++ animation_nodes/nodes/text/timecode.py
@@ -9,7 +9,7 @@ class TimecodeGeneratorNode(bpy.types.Node, AnimationNode):
     def create(self):
         self.newInput("Float", "Frame", "frame")
         socket = self.newInput("Float", "Frame Rate", "frameRate", minValue = 0)
-        socket.value = bpy.context.scene.render.fps
+        socket.value = bpy.context.scene.render.fps / bpy.context.scene.render.fps_base
         self.newOutput("Text", "Timecode", "timecode")
 
     def execute(self, frame, frameRate):